--- a/lib/ops/opsservice/configure.go
+++ b/lib/ops/opsservice/configure.go
@@ -1444,9 +1444,9 @@ func (s *site) addClusterConfig(config clusterconfig.Interface, overrideArgs map
 	if globalConfig.HighAvailability != nil && *globalConfig.HighAvailability {
 		args = append(args, "--high-availability")
 	}
-	if globalConfig.FlannelBackend != "" {
+	if globalConfig.FlannelBackend != nil {
 		args = append(args,
-			fmt.Sprintf("--flannel-backend=%v", globalConfig.FlannelBackend))
+			fmt.Sprintf("--flannel-backend=%v", *globalConfig.FlannelBackend))
 	}
 	return args
 }