## Description
<!--Required. Provide high-level overview of what the change is for.-->
I guess this is similar to the HA reconfiguration [fix](https://github.com/gravitational/gravity/pull/2573) - enable cluster configuration updates on an existing cluster (when flannel backend is not updated).

Configuration file used:
```yaml
kind: ClusterConfiguration
version: v1
spec:
  global:
    cloudProvider: generic
    highAvailability: true
```

Before the change:
```
$ gravity resource create -f haconfig-active.yaml --confirm
[ERROR]: unsupported flannel backend was specified:
```

After the change:
```
$ gravity resource create -f haconfig-active.yaml --confirm
Wed Aug 11 12:44:01 UTC	Deploying agents on cluster nodes
Deployed agent on vm-3 (172.17.0.4)
Deployed agent on vm-2 (172.17.0.3)
Deployed agent on vm-1 (172.17.0.2)
...
```
<details><summary><code>Logs</code></summary>

```
Executing "/update-config" locally
Wed Aug 11 12:44:12 UTC	Update runtime configuration
Executing "/masters/vm-1/stepdown" locally
Wed Aug 11 12:44:15 UTC	Step down "vm-1" as Kubernetes leader
Disable leader elections on vm-1
Waiting for new leader election
Executing "/masters/vm-1/drain" locally
Wed Aug 11 12:44:17 UTC	Drain node "vm-1"
	Still executing "/masters/vm-1/drain" locally (10 seconds elapsed)
Executing "/masters/vm-1/restart" locally
Wed Aug 11 12:44:33 UTC	Restart container on node "vm-1"
	Still executing "/masters/vm-1/restart" locally (10 seconds elapsed)
	Still executing "/masters/vm-1/restart" locally (20 seconds elapsed)
	Still executing "/masters/vm-1/restart" locally (30 seconds elapsed)
	Still executing "/masters/vm-1/restart" locally (40 seconds elapsed)
	Still executing "/masters/vm-1/restart" locally (50 seconds elapsed)
Executing "/masters/vm-1/elect" locally
Wed Aug 11 12:45:23 UTC	Make node "vm-1" Kubernetes leader
Disable leader elections on vm-3
Disable leader elections on vm-2
Enable leader elections on vm-1
Waiting for new leader election
Executing "/masters/vm-1/taint" locally
Wed Aug 11 12:45:26 UTC	Taint node "vm-1"
	Still executing "/masters/vm-1/taint" locally (10 seconds elapsed)
Executing "/masters/vm-1/uncordon" locally
Wed Aug 11 12:45:40 UTC	Uncordon node "vm-1"
Executing "/masters/vm-1/endpoints" locally
Wed Aug 11 12:45:41 UTC	Wait for endpoints on node "vm-1"
Executing "/masters/vm-1/untaint" locally
Wed Aug 11 12:45:42 UTC	Remove taint from node "vm-1"
Executing "/masters/vm-3/drain" on remote node vm-3
Wed Aug 11 12:45:43 UTC	Drain node "vm-3"
	Still executing "/masters/vm-3/drain" on remote node vm-3 (10 seconds elapsed)
Executing "/masters/vm-3/restart" on remote node vm-3
Wed Aug 11 12:45:57 UTC	Restart container on node "vm-3"
	Still executing "/masters/vm-3/restart" on remote node vm-3 (10 seconds elapsed)
	Still executing "/masters/vm-3/restart" on remote node vm-3 (20 seconds elapsed)
	Still executing "/masters/vm-3/restart" on remote node vm-3 (30 seconds elapsed)
Executing "/masters/vm-3/taint" on remote node vm-3
Wed Aug 11 12:46:33 UTC	Taint node "vm-3"
Executing "/masters/vm-3/uncordon" on remote node vm-3
Wed Aug 11 12:46:35 UTC	Uncordon node "vm-3"
Executing "/masters/vm-3/endpoints" on remote node vm-3
Wed Aug 11 12:46:38 UTC	Wait for endpoints on node "vm-3"
Executing "/masters/vm-3/untaint" on remote node vm-3
Wed Aug 11 12:46:41 UTC	Remove taint from node "vm-3"
Executing "/masters/vm-3/enable-elections" on remote node vm-3
Wed Aug 11 12:46:44 UTC	Enable leader election on node "vm-3"
Executing "/masters/vm-2/drain" on remote node vm-2
Wed Aug 11 12:46:47 UTC	Drain node "vm-2"
Executing "/masters/vm-2/restart" on remote node vm-2
Wed Aug 11 12:46:51 UTC	Restart container on node "vm-2"
	Still executing "/masters/vm-2/restart" on remote node vm-2 (10 seconds elapsed)
Executing "/masters/vm-2/taint" on remote node vm-2
Wed Aug 11 12:47:07 UTC	Taint node "vm-2"
Executing "/masters/vm-2/uncordon" on remote node vm-2
Wed Aug 11 12:47:09 UTC	Uncordon node "vm-2"
Executing "/masters/vm-2/endpoints" on remote node vm-2
Wed Aug 11 12:47:12 UTC	Wait for endpoints on node "vm-2"
Executing "/masters/vm-2/untaint" on remote node vm-2
Wed Aug 11 12:47:14 UTC	Remove taint from node "vm-2"
Executing "/masters/vm-2/enable-elections" on remote node vm-2
Wed Aug 11 12:47:18 UTC	Enable leader election on node "vm-2"
Wed Aug 11 12:47:22 UTC	Operation has completed
operation(update configuration(db0c9be5-5c81-4c45-8525-af7d64034c0a), cluster=test, created=2021-08-11 12:44) finished in 3 minutes
```
</details>

## Type of change
<!--Required. Keep only those that apply.-->

* Bug fix (non-breaking change which fixes an issue)
* Regression fix (non-breaking change which fixes a regression)
* This change has a user-facing impact

## TODOs
<!--Required. Keep only those that apply and check them off as they get completed.-->

- [x] Self-review the change
- [ ] Write tests
- [x] Perform manual testing
- [ ] Address review feedback

## Implementation
<!--Optional. Add any relevant implementation details that might help the reviewers.-->

## Performance/Scaling
<!--Optional. Add any relevant details on how this PR reacts when scaled to 1k nodes, and any additional scaling considerations for the reviewers.-->

## Testing done
<!--Required. Explain what kind of testing these changes underwent.-->
1. Updated cluster configuration with the configuration file from above.
2. Attempt to updated flannel backend from an invalid value (explicitly empty and unsupported value).
3. Update flannel backend using an explicitly set value.

## Additional information
<!--Optional. Anything else that may be relevant.-->